+++ app/src/main/java/com/pedroid/qrcodecompose/androidapp/core/presentation/QRAppActionExtensions.kt
@@ -2,6 +2,8 @@ package com.pedroid.qrcodecompose.androidapp.core.presentation
 
 import com.pedroid.qrcodecompose.androidapp.core.domain.ActionStatus
 import com.pedroid.qrcodecompose.androidapp.core.domain.QRAppActions
+import com.pedroid.qrcodecompose.androidapp.core.presentation.composables.TemporaryMessageData
+import com.pedroid.qrcodecompose.androidapp.core.presentation.composables.TemporaryMessageType
 
 fun QRAppActions.asTemporaryMessage(): TemporaryMessageData? {
     return when (this.status) {
