+++ app/src/test/java/com/pedroid/qrcodecompose/androidapp/core/presentation/QRAppActionsTest.kt
@@ -2,6 +2,7 @@ package com.pedroid.qrcodecompose.androidapp.core.presentation
 
 import com.pedroid.qrcodecompose.androidapp.core.domain.ActionStatus
 import com.pedroid.qrcodecompose.androidapp.core.domain.QRAppActions
+import com.pedroid.qrcodecompose.androidapp.core.presentation.composables.TemporaryMessageType
 import org.junit.Assert.assertEquals
 import org.junit.Test
