+++ app/src/test/java/com/pedroid/qrcodecompose/androidapp/features/generate/presentation/GenerateQRCodeViewModelTest.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.SavedStateHandle
 import app.cash.turbine.test
 import com.pedroid.qrcodecompose.androidapp.core.domain.ActionStatus
 import com.pedroid.qrcodecompose.androidapp.core.domain.QRAppActions
-import com.pedroid.qrcodecompose.androidapp.core.presentation.TemporaryMessageData
-import com.pedroid.qrcodecompose.androidapp.core.presentation.TemporaryMessageType
+import com.pedroid.qrcodecompose.androidapp.core.presentation.composables.TemporaryMessageData
+import com.pedroid.qrcodecompose.androidapp.core.presentation.composables.TemporaryMessageType
 import com.pedroid.qrcodecompose.androidapp.core.test.CoroutineDispatcherTestRule
 import com.pedroid.qrcodecompose.androidapp.features.generate.data.QRCodeCustomizationOptions
 import com.pedroid.qrcodecompose.androidapp.features.generate.data.QRCodeGeneratingContent
